+++ tests/testthat/test-fetchPrioritizations.R
@@ -2,7 +2,7 @@ context("fetch-prioritizations")
 
 with_mock_api({
   test_that("We can fetch PSNU prioritizations from DATIM", {
-    prios <- fetchPrioritizationTable(data.frame(psnu_uid = c("uXwFHXCPYgj")),
+    prios <- fetchPrioritizationTable(c("uXwFHXCPYgj"),
                                       "2022",
                                       d2_session = training)
     expect_true(is.data.frame(prios))