+++ tests/testthat/test-unpack-COP21-datapack.R
@@ -8,7 +8,7 @@ d <-
     submission_path = test_sheet("COP21_DP_random_no_psnuxim.xlsx"),
     tool = "Data Pack",
     country_uids = NULL,
-    cop_year = NULL,
+    cop_year = 2021,
     load_sheets = TRUE,
     d2_session = training)